move ddl worker ut to the sink
