+++ downstreamadapter/sink/cloudstorage/sink_test.go
@@ -17,6 +17,8 @@ import (
 	"context"
 	"fmt"
 	"net/url"
+	"os"
+	"path"
 	"sync/atomic"
 	"testing"
 	"time"
@@ -26,10 +28,15 @@ import (
 	commonEvent "github.com/pingcap/ticdc/pkg/common/event"
 	"github.com/pingcap/ticdc/pkg/config"
 	"github.com/pingcap/ticdc/pkg/pdutil"
+	"github.com/pingcap/ticdc/pkg/util"
+	timodel "github.com/pingcap/tidb/pkg/meta/model"
+	pmodel "github.com/pingcap/tidb/pkg/parser/model"
+	"github.com/pingcap/tidb/pkg/parser/mysql"
+	"github.com/pingcap/tidb/pkg/types"
 	"github.com/stretchr/testify/require"
 )
 
-func newCloudStorageSinkForTest(parentDir string) (*sink, error) {
+func newSinkForTest(parentDir string, cleanUpJobs []func()) (*sink, error) {
 	ctx := context.Background()
 	mockPDClock := pdutil.NewClock4Test()
 	appcontext.SetService(appcontext.DefaultPDClock, mockPDClock)
@@ -41,17 +48,21 @@ func newCloudStorageSinkForTest(parentDir string) (*sink, error) {
 	if err != nil {
 		return nil, err
 	}
-	sink, err := New(ctx, changefeedID, sinkURI, sinkConfig, nil)
+	sink, err := New(ctx, changefeedID, sinkURI, sinkConfig, cleanUpJobs)
 	if err != nil {
 		return nil, err
 	}
-	go sink.Run(ctx)
 	return sink, nil
+
 }
 
-func TestCloudStorageSinkBasicFunctionality(t *testing.T) {
-	sink, err := newCloudStorageSinkForTest(t.TempDir())
+func TestBasicFunctionality(t *testing.T) {
+	cloudStorageSink, err := newSinkForTest(t.TempDir(), nil)
 	require.NoError(t, err)
+
+	ctx, cancel := context.WithCancel(context.Background())
+	defer cancel()
+	go cloudStorageSink.Run(ctx)
 
 	var count atomic.Int64
 
@@ -106,14 +117,121 @@ func TestCloudStorageSinkBasicFunctionality(t *testing.T) {
 	}
 	dmlEvent.TableInfoVersion = 1
 
-	err = sink.WriteBlockEvent(ddlEvent)
+	err = cloudStorageSink.WriteBlockEvent(ddlEvent)
 	require.NoError(t, err)
 
-	sink.AddDMLEvent(dmlEvent)
+	cloudStorageSink.AddDMLEvent(dmlEvent)
 
 	time.Sleep(5 * time.Second)
 
 	ddlEvent2.PostFlush()
 
 	require.Equal(t, count.Load(), int64(3))
+}
+
+func TestWriteDDLEvent(t *testing.T) {
+	parentDir := t.TempDir()
+	cloudStorageSink, err := newSinkForTest(parentDir, nil)
+	require.NoError(t, err)
+
+	ctx, cancel := context.WithCancel(context.Background())
+	defer cancel()
+	go cloudStorageSink.Run(ctx)
+
+	tableInfo := common.WrapTableInfo("test", &timodel.TableInfo{
+		ID:   20,
+		Name: pmodel.NewCIStr("table1"),
+		Columns: []*timodel.ColumnInfo{
+			{
+				Name:      pmodel.NewCIStr("col1"),
+				FieldType: *types.NewFieldType(mysql.TypeLong),
+			},
+			{
+				Name:      pmodel.NewCIStr("col2"),
+				FieldType: *types.NewFieldType(mysql.TypeVarchar),
+			},
+		},
+	})
+	ddlEvent := &commonEvent.DDLEvent{
+		Query:      "alter table test.table1 add col2 varchar(64)",
+		Type:       byte(timodel.ActionAddColumn),
+		SchemaName: "test",
+		TableName:  "table1",
+		FinishedTs: 100,
+		TableInfo:  tableInfo,
+	}
+
+	tableDir := path.Join(parentDir, "test/table1/meta/")
+	err = cloudStorageSink.WriteBlockEvent(ddlEvent)
+	require.NoError(t, err)
+
+	tableSchema, err := os.ReadFile(path.Join(tableDir, "schema_100_4192708364.json"))
+	require.NoError(t, err)
+	require.JSONEq(t, `{
+		"Table": "table1",
+		"Schema": "test",
+		"Version": 1,
+		"TableVersion": 100,
+		"Query": "alter table test.table1 add col2 varchar(64)",
+		"Type": 5,
+		"TableColumns": [
+			{
+				"ColumnName": "col1",
+				"ColumnType": "INT",
+				"ColumnPrecision": "11"
+			},
+			{
+				"ColumnName": "col2",
+				"ColumnType": "VARCHAR",
+				"ColumnPrecision": "5"
+			}
+		],
+		"TableColumnsTotal": 2
+	}`, string(tableSchema))
+}
+
+func TestWriteCheckpointEvent(t *testing.T) {
+	parentDir := t.TempDir()
+	cloudStorageSink, err := newSinkForTest(parentDir, nil)
+	require.NoError(t, err)
+
+	ctx, cancel := context.WithCancel(context.Background())
+	defer cancel()
+	go cloudStorageSink.Run(ctx)
+
+	time.Sleep(3 * time.Second)
+	cloudStorageSink.AddCheckpointTs(100)
+	metadata, err := os.ReadFile(path.Join(parentDir, "metadata"))
+	require.NoError(t, err)
+	require.JSONEq(t, `{"checkpoint-ts":100}`, string(metadata))
+}
+
+func TestCleanupExpiredFiles(t *testing.T) {
+	parentDir := t.TempDir()
+	uri := fmt.Sprintf("file:///%s?protocol=csv", parentDir)
+	sinkURI, err := url.Parse(uri)
+	require.NoError(t, err)
+	replicaConfig := config.GetDefaultReplicaConfig()
+	replicaConfig.Sink.CloudStorageConfig = &config.CloudStorageConfig{
+		FileExpirationDays:  util.AddressOf(1),
+		FileCleanupCronSpec: util.AddressOf("* * * * * *"),
+	}
+	err = replicaConfig.ValidateAndAdjust(sinkURI)
+	require.NoError(t, err)
+
+	var count atomic.Int64
+	cleanupJobs := []func(){
+		func() {
+			count.Add(1)
+		},
+	}
+
+	ctx, cancel := context.WithCancel(context.Background())
+	defer cancel()
+	cloudStorageSink, err := newSinkForTest(parentDir, cleanupJobs)
+	go cloudStorageSink.Run(ctx)
+	require.NoError(t, err)
+
+	time.Sleep(5 * time.Second)
+	require.LessOrEqual(t, int64(1), count.Load())
 }
